+++ packages/api/src/env.js
@@ -1,6 +1,5 @@
-/* global MAGIC_SECRET_KEY FAUNA_ENDPOINT FAUNA_KEY SALT CLUSTER_BASIC_AUTH_TOKEN CLUSTER_API_URL SENTRY_DSN SENTRY_RELEASE DANGEROUSLY_BYPASS_MAGIC_AUTH */
+/* global MAGIC_SECRET_KEY SALT CLUSTER_BASIC_AUTH_TOKEN CLUSTER_API_URL SENTRY_DSN SENTRY_RELEASE DANGEROUSLY_BYPASS_MAGIC_AUTH PG_REST_URL PG_REST_JWT */
 /* global S3_BUCKET_ENDPOINT S3_BUCKET_NAME S3_BUCKET_REGION S3_ACCESS_KEY_ID S3_SECRET_ACCESS_KEY_ID ENV MAINTENANCE_MODE VERSION COMMITHASH BRANCH */
-/* global DATABASE PG_REST_URL PG_REST_JWT */
 import Toucan from 'toucan-js'
 import { S3Client } from '@aws-sdk/client-s3'
 import { Magic } from '@magic-sdk/admin'
@@ -50,20 +49,10 @@ export function envAll (_, env, event) {
     env.DANGEROUSLY_BYPASS_MAGIC_AUTH = DANGEROUSLY_BYPASS_MAGIC_AUTH
   }
 
-  if (env.DATABASE === 'fauna' ||
-    (typeof DATABASE !== 'undefined' && DATABASE === 'fauna') ||
-    (!env.DATABASE && typeof DATABASE === 'undefined')) {
-    env.db = new DBClient({
-      endpoint: env.FAUNA_ENDPOINT || (typeof FAUNA_ENDPOINT === 'undefined' ? undefined : FAUNA_ENDPOINT),
-      token: env.FAUNA_KEY || FAUNA_KEY
-    })
-  } else {
-    env.db = new DBClient({
-      endpoint: env.PG_REST_URL || (typeof PG_REST_URL === 'undefined' ? undefined : PG_REST_URL),
-      token: env.PG_REST_JWT || PG_REST_JWT,
-      postgres: true
-    })
-  }
+  env.db = new DBClient({
+    endpoint: env.PG_REST_URL || PG_REST_URL,
+    token: env.PG_REST_JWT || PG_REST_JWT
+  })
 
   env.SALT = env.SALT || SALT
   env.MODE = env.MAINTENANCE_MODE || (typeof MAINTENANCE_MODE === 'undefined' ? DEFAULT_MODE : MAINTENANCE_MODE)