+++ packages/db/index.js
@@ -1,31 +1,81 @@
-import { gql } from 'graphql-request'
+import { PostgrestClient } from '@supabase/postgrest-js'
 
-import { FaunaClient } from './fauna/client.js'
-import { PostgresClient } from './postgres/client.js'
+import { normalizeUpload, normalizeContent, normalizePins, normalizePaPinRequest } from './utils.js'
+import { DBError } from './errors.js'
+import {
+  getUserMetrics,
+  getUploadMetrics,
+  getPinMetrics,
+  getPinStatusMetrics,
+  getContentMetrics,
+  getPinBytesMetrics
+} from './metrics.js'
 
-const ENDPOINT = 'https://graphql.fauna.com/graphql'
+const uploadQuery = `
+        _id:id::text,
+        type,
+        name,
+        created:inserted_at,
+        updated:updated_at,
+        content(cid, dagSize:dag_size, pins:pin(status, updated:updated_at, location:pin_location(_id:id, peerId:peer_id, peerName:peer_name, region)))
+      `
 
-// TODO: Deprecate with fauna
-export { gql }
+const PAPinRequestTableName = 'pa_pin_request'
+const pinRequestSelect = `
+  _id:id::text,
+  requestedCid:requested_cid,
+  contentCid:content_cid,
+  authKey:auth_key_id,
+  name,
+  created:inserted_at,
+  updated:updated_at,
+  content(cid, dagSize:dag_size, pins:pin(status, updated:updated_at, location:pin_location(_id:id, peerId:peer_id, peerName:peer_name, region)))  `
+
+/**
+ * @typedef {import('./postgres/pg-rest-api-types').definitions} definitions
+ * @typedef {import('@supabase/postgrest-js').PostgrestError} PostgrestError
+ */
 
 export class DBClient {
-  constructor ({ endpoint = ENDPOINT, token, postgres = false }) {
-    if (postgres) {
-      this._client = new PostgresClient({ endpoint, token })
-      this._isPostgres = true
-    } else {
-      this._client = new FaunaClient({ endpoint, token })
-    }
+  constructor ({ endpoint, token }) {
+    this._client = new PostgrestClient(endpoint, {
+      headers: {
+        Authorization: `Bearer ${token}`,
+        accept: '*/*'
+      }
+    })
   }
 
   /**
    * Upsert user.
    *
    * @param {import('./db-client-types').UpsertUserInput} user
-   * @return {import('./db-client-types').UpsertUserOutput}
+   * @return {Promise<import('./db-client-types').UpsertUserOutput>}
    */
-  upsertUser (user) {
-    return this._client.upsertUser(user)
+  async upsertUser (user) {
+    /** @type {{ data: definitions['user'], error: PostgrestError }} */
+    const { data, error } = await this._client
+      .from('user')
+      .upsert({
+        id: user.id,
+        name: user.name,
+        picture: user.picture,
+        email: user.email,
+        issuer: user.issuer,
+        github: user.github,
+        public_address: user.publicAddress
+      }, {
+        onConflict: 'issuer'
+      })
+      .single()
+
+    if (error) {
+      throw new DBError(error)
+    }
+
+    return {
+      issuer: data.issuer
+    }
   }
 
   /**
@@ -34,8 +84,28 @@ export class DBClient {
    * @param {string} issuer
    * @return {Promise<import('./db-client-types').UserOutput>}
    */
-  getUser (issuer) {
-    return this._client.getUser(issuer)
+  async getUser (issuer) {
+    /** @type {{ data: import('./db-client-types').UserOutput, error: PostgrestError }} */
+    const { data, error } = await this._client
+      .from('user')
+      .select(`
+        _id:id::text,
+        issuer,
+        name,
+        email,
+        github,
+        publicAddress:public_address,
+        created:inserted_at,
+        updated:updated_at
+      `)
+      .eq('issuer', issuer)
+      .single()
+
+    if (error) {
+      throw new DBError(error)
+    }
+
+    return data
   }
 
   /**
@@ -44,8 +114,15 @@ export class DBClient {
    * @param {number} userId
    * @returns {Promise<number>}
    */
-  getUsedStorage (userId) {
-    return this._client.getUsedStorage(userId)
+  async getUsedStorage (userId) {
+    /** @type {{ data: string, error: PostgrestError }} */
+    const { data, error } = await this._client.rpc('user_used_storage', { query_user_id: userId }).single()
+
+    if (error) {
+      throw new DBError(error)
+    }
+
+    return data || 0 // No uploads for the user
   }
 
   /**
@@ -54,8 +131,41 @@ export class DBClient {
    * @param {import('./db-client-types').CreateUploadInput} data
    * @returns {Promise<import('./db-client-types').CreateUploadOutput>}
    */
-  createUpload (data) {
-    return this._client.createUpload(data)
+  async createUpload (data) {
+    // TODO(paolo): refactor this to be createContent, and optionally an upload
+    const now = new Date().toISOString()
+    /** @type {{ data: string, error: PostgrestError }} */
+    const { data: uploadResponse, error } = await this._client.rpc('create_upload', {
+      data: {
+        user_id: data.user,
+        auth_key_id: data.authKey,
+        content_cid: data.contentCid,
+        source_cid: data.sourceCid,
+        type: data.type,
+        name: data.name,
+        dag_size: data.dagSize,
+        inserted_at: data.created || now,
+        updated_at: data.updated || now,
+        pins: data.pins.map(pin => ({
+          status: pin.status,
+          location: {
+            peer_id: pin.location.peerId,
+            peer_name: pin.location.peerName,
+            region: pin.location.region
+          }
+        })),
+        backup_urls: data.backupUrls
+      }
+    }).single()
+
+    if (error) {
+      throw new DBError(error)
+    }
+
+    return {
+      _id: uploadResponse,
+      cid: data.sourceCid
+    }
   }
 
   /**
@@ -65,8 +175,26 @@ export class DBClient {
    * @param {number} userId
    * @returns {Promise<import('./db-client-types').UploadItemOutput>}
    */
-  getUpload (cid, userId) {
-    return this._client.getUpload(cid, userId)
+  async getUpload (cid, userId) {
+    /** @type {{ data: import('./db-client-types').UploadItem, error: PostgrestError }} */
+    const { data: upload, error } = await this._client
+      .from('upload')
+      .select(uploadQuery)
+      .eq('source_cid', cid)
+      .eq('user_id', userId)
+      .is('deleted_at', null)
+      .single()
+
+    if (error) {
+      throw new DBError(error)
+    }
+
+    const deals = await this.getDeals(cid)
+
+    return {
+      ...normalizeUpload(upload),
+      deals
+    }
   }
 
   /**
@@ -76,8 +204,41 @@ export class DBClient {
    * @param {import('./db-client-types').ListUploadsOptions} [opts]
    * @returns {Promise<Array<import('./db-client-types').UploadItemOutput>>}
    */
-  listUploads (userId, opts = {}) {
-    return this._client.listUploads(userId, opts)
+  async listUploads (userId, opts = {}) {
+    let query = this._client
+      .from('upload')
+      .select(uploadQuery)
+      .eq('user_id', userId)
+      .is('deleted_at', null)
+      .limit(opts.size || 10)
+      .order(
+        opts.sortBy === 'Name' ? 'name' : 'inserted_at',
+        { ascending: opts.sortOrder === 'Asc' }
+      )
+
+    if (opts.before) {
+      query = query.lt('inserted_at', opts.before)
+    }
+
+    if (opts.after) {
+      query = query.gte('inserted_at', opts.after)
+    }
+
+    /** @type {{ data: Array<import('./db-client-types').UploadItem>, error: Error }} */
+    const { data: uploads, error } = await query
+
+    if (error) {
+      throw new DBError(error)
+    }
+
+    // Get deals
+    const cids = uploads?.map((u) => u.content.cid)
+    const deals = await this.getDealsForCids(cids)
+
+    return uploads?.map((u) => ({
+      ...normalizeUpload(u),
+      deals: deals[u.content.cid] || []
+    }))
   }
 
   /**
@@ -87,8 +248,25 @@ export class DBClient {
    * @param {string} cid
    * @param {string} name
    */
-  renameUpload (userId, cid, name) {
-    return this._client.renameUpload(userId, cid, name)
+  async renameUpload (userId, cid, name) {
+    /** @type {{ data: import('./db-client-types').UploadItem, error: PostgrestError }} */
+    const { data, error } = await this._client
+      .from('upload')
+      .update({ name })
+      .match({
+        user_id: userId,
+        source_cid: cid
+      })
+      .is('deleted_at', null)
+      .single()
+
+    if (error) {
+      throw new DBError(error)
+    }
+
+    return {
+      name: data.name
+    }
   }
 
   /**
@@ -97,8 +275,27 @@ export class DBClient {
    * @param {number} userId
    * @param {string} cid
    */
-  deleteUpload (userId, cid) {
-    return this._client.deleteUpload(userId, cid)
+  async deleteUpload (userId, cid) {
+    /** @type {{ data: import('./db-client-types').UploadItem, error: PostgrestError }} */
+    const { data, error } = await this._client
+      .from('upload')
+      .update({
+        deleted_at: new Date().toISOString()
+      })
+      .match({
+        source_cid: cid,
+        user_id: userId
+      })
+      .is('deleted_at', null)
+      .single()
+
+    if (error) {
+      throw new DBError(error)
+    }
+
+    return {
+      _id: data.id
+    }
   }
 
   /**
@@ -107,8 +304,32 @@ export class DBClient {
    * @param {string} cid
    * @returns {Promise<import('./db-client-types').ContentItemOutput>}
    */
-  getStatus (cid) {
-    return this._client.getStatus(cid)
+  async getStatus (cid) {
+    /** @type {{ data: import('./db-client-types').ContentItem, error: PostgrestError }} */
+    const { data, error } = await this._client
+      .from('content')
+      .select(`
+        cid,
+        dagSize:dag_size,
+        created:inserted_at,
+        pins:pin(status, updated:updated_at, location:pin_location(peerId:peer_id, peerName:peer_name, region))
+      `)
+      .match({ cid })
+      .single()
+
+    if (error) {
+      throw new DBError(error)
+    }
+
+    if (!data) {
+      return
+    }
+
+    const deals = await this.getDeals(cid)
+    return {
+      ...normalizeContent(data),
+      deals
+    }
   }
 
   /**
@@ -117,19 +338,53 @@ export class DBClient {
    * @param {number} uploadId
    * @return {Promise<Array<import('./db-client-types').BackupOutput>>}
    */
-  getBackups (uploadId) {
-    return this._client.getBackups(uploadId)
+  async getBackups (uploadId) {
+    /** @type {{ data: Array<definitions['backup']>, error: PostgrestError }} */
+    const { data: backups, error } = await this._client
+      .from('backup')
+      .select(`
+        _id:id::text,
+        created:inserted_at,
+        uploadId:upload_id::text,
+        url
+      `)
+      .match({ upload_id: uploadId })
+
+    if (error) {
+      throw new DBError(error)
+    }
+
+    return backups
   }
 
   /**
-   * Upsert pin.
+   * Upsert pin of a given cid.
    *
    * @param {string} cid
    * @param {import('./db-client-types').PinUpsertInput} pin
    * @return {Promise<number>}
    */
-  upsertPin (cid, pin) {
-    return this._client.upsertPin(cid, pin)
+  async upsertPin (cid, pin) {
+    /** @type {{ data: number, error: PostgrestError }} */
+    const { data: pinId, error } = await this._client.rpc('upsert_pin', {
+      data: {
+        content_cid: cid,
+        pin: {
+          status: pin.status,
+          location: {
+            peer_id: pin.location.peerId,
+            peer_name: pin.location.peerName,
+            region: pin.location.region
+          }
+        }
+      }
+    }).single()
+
+    if (error) {
+      throw new DBError(error)
+    }
+
+    return pinId
   }
 
   /**
@@ -137,8 +392,14 @@ export class DBClient {
    *
    * @param {Array<import('./db-client-types').PinUpsertInput>} pins
    */
-  upsertPins (pins) {
-    return this._client.upsertPins(pins)
+  async upsertPins (pins) {
+    const { error } = await this._client
+      .from('pin')
+      .upsert(pins, { count: 'exact', returning: 'minimal' })
+
+    if (error) {
+      throw new DBError(error)
+    }
   }
 
   /**
@@ -147,8 +408,24 @@ export class DBClient {
    * @param {string} cid
    * @return {Promise<Array<import('./db-client-types').PinItemOutput>>}
    */
-  getPins (cid) {
-    return this._client.getPins(cid)
+  async getPins (cid) {
+    /** @type {{ data: Array<import('./db-client-types').PinItem>, error: PostgrestError }} */
+    const { data: pins, error } = await this._client
+      .from('pin')
+      .select(`
+        _id:id::text,
+        status,
+        created:inserted_at,
+        updated:updated_at,
+        location:pin_location(id::text, peerId:peer_id, peerName:peer_name, region)
+      `)
+      .match({ content_cid: cid })
+
+    if (error) {
+      throw new DBError(error)
+    }
+
+    return normalizePins(pins)
   }
 
   /**
@@ -156,10 +433,24 @@ export class DBClient {
    *
    * @param {Object} [options]
    * @param {number} [options.size = 600]
-   * @return {Promise<Array<import('../db-client-types').PinRequestItemOutput>>}
+   * @return {Promise<Array<import('./db-client-types').PinRequestItemOutput>>}
    */
-  getPinRequests () {
-    return this._client.getPinRequests()
+  async getPinRequests ({ size = 600 } = {}) {
+    /** @type {{ data: Array<import('./db-client-types').PinRequestItemOutput>, error: PostgrestError }} */
+    const { data: pinReqs, error } = await this._client
+      .from('pin_request')
+      .select(`
+        _id:id::text,
+        cid:content_cid,
+        created:inserted_at
+      `)
+      .limit(size)
+
+    if (error) {
+      throw new DBError(error)
+    }
+
+    return pinReqs
   }
 
   /**
@@ -168,8 +459,16 @@ export class DBClient {
    * @param {Array<number>} ids
    * @return {Promise<void>}
    */
-  deletePinRequests (ids) {
-    return this._client.deletePinRequests(ids)
+  async deletePinRequests (ids) {
+    /** @type {{ error: PostgrestError }} */
+    const { error } = await this._client
+      .from('pin_request')
+      .delete()
+      .in('id', ids)
+
+    if (error) {
+      throw new DBError(error)
+    }
   }
 
   /**
@@ -177,8 +476,20 @@ export class DBClient {
    *
    * @param {Array<number>} pinSyncRequests
    */
-  createPinSyncRequests (pinSyncRequests) {
-    return this._client.createPinSyncRequests(pinSyncRequests)
+  async createPinSyncRequests (pinSyncRequests) {
+    /** @type {{ error: PostgrestError }} */
+    const { error } = await this._client
+      .from('pin_sync_request')
+      .upsert(pinSyncRequests.map(psr => ({
+        pin_id: psr,
+        inserted_at: new Date().toISOString()
+      })), {
+        onConflict: 'pin_id'
+      })
+
+    if (error) {
+      throw new DBError(error)
+    }
   }
 
   /**
@@ -186,11 +497,40 @@ export class DBClient {
    *
    * @param {Object} [options]
    * @param {string} [options.to]
-   * @param {string} [options.afer]
-   * @return {Promise<Array<import('../db-client-types').PinSyncRequestOutput>>}
+   * @param {string} [options.after]
+   * @param {number} [options.size]
+   * @return {Promise<import('./db-client-types').PinSyncRequestOutput>}
    */
-  getPinSyncRequests (options) {
-    return this._client.getPinSyncRequests(options)
+  async getPinSyncRequests ({ to, after, size }) {
+    let query = this._client
+      .from('pin_sync_request')
+      .select(`
+        _id:id::text,
+        pin:pin(_id:id::text, status, contentCid:content_cid, created:inserted_at, location:pin_location(_id:id::text, peerId:peer_id, peerName:peer_name, region))
+      `)
+      .order(
+        'inserted_at',
+        { ascending: true }
+      )
+      .limit(size)
+
+    if (to) {
+      query = query.lt('inserted_at', to)
+    }
+    if (after) {
+      query = query.gte('inserted_at', after)
+    }
+
+    /** @type {{ data: Array<import('./db-client-types').PinSyncRequestItem>, error: PostgrestError }} */
+    const { data: pinSyncReqs, error } = await query
+    if (error) {
+      throw new DBError(error)
+    }
+
+    return {
+      data: pinSyncReqs,
+      after: !!size && pinSyncReqs.length === size && pinSyncReqs[0].pin.created // return after if more
+    }
   }
 
   /**
@@ -199,8 +539,16 @@ export class DBClient {
    * @param {Array<number>} ids
    * @return {Promise<void>}
    */
-  deletePinSyncRequests (ids) {
-    return this._client.deletePinSyncRequests(ids)
+  async deletePinSyncRequests (ids) {
+    /** @type {{ error: PostgrestError }} */
+    const { error } = await this._client
+      .from('pin_sync_request')
+      .delete()
+      .in('id', ids)
+
+    if (error) {
+      throw new DBError(error)
+    }
   }
 
   /**
@@ -209,8 +557,9 @@ export class DBClient {
    * @param {string} cid
    * @return {Promise<import('./db-client-types').Deal[]>}
    */
-  getDeals (cid) {
-    return this._client.getDeals(cid)
+  async getDeals (cid) {
+    const deals = await this.getDealsForCids([cid])
+    return deals[cid] ? deals[cid] : []
   }
 
   /**
@@ -219,8 +568,29 @@ export class DBClient {
    * @param {string[]} cids
    * @return {Promise<Record<string, import('./db-client-types').Deal[]>>}
    */
-  getDealsForCids (cids = []) {
-    return this._client.getDealsForCids(cids)
+  async getDealsForCids (cids = []) {
+    /** @type {{ data: Array<import('./db-client-types').Deal>, error: PostgrestError }} */
+    const { data, error } = await this._client
+      .rpc('find_deals_by_content_cids', {
+        cids
+      })
+
+    if (error) {
+      throw new DBError(error)
+    }
+
+    // TODO: normalize deal by removing deal prefix on dealActivation and dealExpiration
+    const result = {}
+    for (const d of data) {
+      const cid = d.dataCid
+      if (!Array.isArray(result[cid])) {
+        result[cid] = [d]
+      } else {
+        result[cid].push(d)
+      }
+    }
+
+    return result
   }
 
   /**
@@ -229,8 +599,27 @@ export class DBClient {
    * @param {import('./db-client-types').CreateAuthKeyInput} key
    * @return {Promise<import('./db-client-types').CreateAuthKeyOutput>}
    */
-  createKey ({ name, secret, user }) {
-    return this._client.createKey({ name, secret, user })
+  async createKey ({ name, secret, user }) {
+    const now = new Date().toISOString()
+
+    /** @type {{ data: string, error: PostgrestError }} */
+    const { data, error } = await this._client.rpc('create_key', {
+      data: {
+        name,
+        secret,
+        user_id: user,
+        inserted_at: now,
+        updated_at: now
+      }
+    }).single()
+
+    if (error) {
+      throw new DBError(error)
+    }
+
+    return {
+      _id: data
+    }
   }
 
   /**
@@ -240,8 +629,38 @@ export class DBClient {
    * @param {string} secret
    * @return {Promise<import('./db-client-types').AuthKey>}
    */
-  getKey (issuer, secret) {
-    return this._client.getKey(issuer, secret)
+  async getKey (issuer, secret) {
+    /** @type {{ data, error: PostgrestError } */
+    const { data, error } = await this._client
+      .from('user')
+      .select(`
+        _id:id::text,
+        issuer,
+        keys:auth_key_user_id_fkey(_id:id::text, name,secret)
+      `)
+      .match({
+        issuer
+      })
+      .filter('keys.deleted_at', 'is', null)
+      .eq('keys.secret', secret)
+      .single()
+
+    if (error) {
+      throw new DBError(error)
+    }
+
+    if (!data.keys.length) {
+      throw new Error('user has no key with given secret')
+    }
+
+    return {
+      _id: data.keys[0]._id,
+      name: data.keys[0].name,
+      user: {
+        _id: data._id,
+        issuer: data.issuer
+      }
+    }
   }
 
   /**
@@ -250,8 +669,21 @@ export class DBClient {
    * @param {number} userId
    * @return {Promise<Array<import('./db-client-types').AuthKeyItemOutput>>}
    */
-  listKeys (userId) {
-    return this._client.listKeys(userId)
+  async listKeys (userId) {
+    /** @type {{ error: PostgrestError, data: Array<import('./db-client-types').AuthKeyItem> }} */
+    const { data, error } = await this._client.rpc('user_keys_list', { query_user_id: userId })
+
+    if (error) {
+      throw new DBError(error)
+    }
+
+    return data.map(ki => ({
+      _id: ki.id,
+      name: ki.name,
+      secret: ki.secret,
+      created: ki.created,
+      hasUploads: Boolean(ki.uploads)
+    }))
   }
 
   /**
@@ -260,8 +692,26 @@ export class DBClient {
    * @param {number} userId
    * @param {number} keyId
    */
-  deleteKey (userId, keyId) {
-    return this._client.deleteKey(userId, keyId)
+  async deleteKey (userId, keyId) {
+    /** @type {{ data, error: PostgrestError }} */
+    const { data, error } = await this._client
+      .from('auth_key')
+      .update({
+        deleted_at: new Date().toISOString()
+      })
+      .match({
+        id: keyId,
+        user_id: userId
+      })
+      .is('deleted_at', null)
+
+    if (error) {
+      throw new DBError(error)
+    }
+
+    return {
+      _id: data.id
+    }
   }
 
   /**
@@ -269,21 +719,107 @@ export class DBClient {
    *
    * @param {string} key
    */
-  getMetricsValue (key) {
-    return this._client.getMetricsValue(key)
+  async getMetricsValue (key) {
+    let res
+    switch (key) {
+      case 'users_total':
+        res = await getUserMetrics(this._client)
+        return res.total
+      case 'uploads_total':
+        res = await getUploadMetrics(this._client)
+        return res.total
+      case 'content_bytes_total':
+        res = await getContentMetrics(this._client)
+        return res.totalBytes
+      case 'pins_total':
+        res = await getPinMetrics(this._client)
+        return res.total
+      case 'pins_bytes_total':
+        res = await getPinBytesMetrics(this._client)
+        return res.totalBytes
+      case 'pins_status_queued_total':
+      case 'pins_status_pinning_total':
+      case 'pins_status_pinned_total':
+      case 'pins_status_failed_total':
+        res = await getPinStatusMetrics(this._client, key)
+        return res.total
+      default:
+        throw new Error('unknown metric requested')
+    }
   }
 
   /**
-   * @template T
-   * @template V
-   * @param {import('graphql-request').RequestDocument} document
-   * @param {V} variables
-   * @returns {Promise<T>}
+   * Creates a Pin Request.
+   *
+   * @param {import('./db-client-types').PAPinRequestUpsertInput} pinRequest
+   * @return {Promise<import('./db-client-types').PAPinRequestUpsertOutput>}
    */
-  query (document, variables) {
-    if (this._isPostgres) {
-      throw new Error('query is only compatible with fauna')
+  async createPAPinRequest (pinRequest) {
+    /** @type { import('./db-client-types').PAPinRequestUpsertOutput } */
+
+    // TODO is there a better way to avoid 2 queries?
+    // ie. before insert trigger https://dba.stackexchange.com/questions/27178/handling-error-of-foreign-key
+    /** @type {{data: {cid: string}}} */
+    const { data: content } = await this._client
+      .from('content')
+      .select('cid')
+      .eq('cid', pinRequest.requestedCid)
+      .single()
+
+    const toInsert = {
+      requested_cid: pinRequest.requestedCid,
+      auth_key_id: pinRequest.authKey,
+      name: pinRequest.name
     }
-    return this._client.query(document, variables)
+
+    // If content already exists updated foreigh key
+    if (content?.cid) {
+      toInsert.content_cid = content.cid
+    }
+
+    /** @type {{data: import('./db-client-types').PAPinRequestItem, error: PostgrestError }} */
+    const { data, error } = await this._client
+      .from(PAPinRequestTableName)
+      .insert(toInsert)
+      .select(pinRequestSelect)
+      .single()
+
+    if (error) {
+      throw new DBError(error)
+    }
+
+    return normalizePaPinRequest(data)
+  }
+
+  /**
+   * Get a Pin Request by id
+   *
+   * @param {string} pinRequestId
+   * @return {Promise<import('./db-client-types').PAPinRequestUpsertOutput>}
+   */
+  async getPAPinRequest (pinRequestId) {
+    /** @type {{data: import('./db-client-types').PAPinRequestItem, error: PostgrestError }} */
+    const { data, error } = await this._client
+      .from(PAPinRequestTableName)
+      .select(pinRequestSelect)
+      .eq('id', pinRequestId)
+      .single()
+
+    if (error) {
+      throw new DBError(error)
+    }
+
+    return normalizePaPinRequest(data)
+  }
+
+  /**
+   * Get pin requests for a specific auth key.
+   *
+   * @param {string} authKey
+   * @param {object} opt
+   * @return {Promise<Array<import('./db-client-types').PAPinRequestUpsertOutput>>}
+   */
+  async listPAPinRequests (authKey, opt) {
+    throw new Error('Not implemented')
   }
 }